issue 29: add doxygen to crc3x.h
I thought this would be a good opportunity to rename the libcrcxxx library to libcrc3x.

Spruced up the documentation a bit.
