+++ .travis.yml
@@ -30,9 +30,9 @@ script:
   - echo 'extern void crcx_init(); int main() { crcx_init(); return 0; }' > foo.c
   - gcc -o foo foo.c `pkg-config --cflags --libs crcx`
   - # ensure a simple application can compile and link to libcrcxxx
-  - echo '#include <crcx/crcxxx.h>' > foo.cpp
-  - echo 'using namespace ::crcx; int main() { using Crcx = Crc<8,uint8_t,7>; Crcx crc(0,0,false,false); return 0; }' >> foo.cpp
-  - g++ -o foo foo.cpp `pkg-config --cflags --libs crcxxx`
+  - echo '#include <crc3x/crc3x.h>' > foo.cpp
+  - echo 'using namespace ::crc3x; int main() { using Crc3x = Crc<8,uint8_t,7>; Crc3x crc(0,0,false,false); return 0; }' >> foo.cpp
+  - g++ -o foo foo.cpp `pkg-config --cflags --libs crc3x`
 
 after_success:
   - bash <(curl -s https://codecov.io/bash)
