+++ configure.ac
@@ -102,7 +102,7 @@ fi
 fi
 AM_CONDITIONAL([HAVE_DOXYGEN], [test -n "$DOXYGEN"])
 AM_COND_IF([HAVE_DOXYGEN], [AC_CONFIG_FILES([docs/Doxyfile])])
-AS_IF([test "x$have_cxx" = "xyes"], [AC_CONFIG_FILES([src/crcxxx.pc])])
+AS_IF([test "x$have_cxx" = "xyes"], [AC_CONFIG_FILES([src/crc3x.pc])])
 
 AC_CONFIG_FILES([
 	Makefile
