+++ src/crc3x.cpp
@@ -1,0 +1,10 @@
+#include "crc3x/crc3x.h"
+
+using namespace crc3x;
+// using namespace std;
+
+namespace crc3x {
+
+void empty() {}
+
+} // namespace crc3x
