--- src/crcxxx.cpp
@@ -1,10 +1,0 @@
-#include "crcx/crcxxx.h"
-
-using namespace crcx;
-// using namespace std;
-
-namespace crcx {
-
-void empty() {}
-
-} /* namespace crcx */
